RAC-1050: Move ProgressCell to common
