+++ src/Akeneo/Platform/Job/front/process-tracker/src/feature/components/common/ProgressCell.test.tsx
