+++ src/Akeneo/Platform/Job/front/process-tracker/src/feature/components/common/index.ts
@@ -1,2 +1,3 @@
 export * from './JobExecutionStatus';
+export * from './ProgressCell';
 export * from './StopJobAction';
