+++ src/Akeneo/Platform/Job/front/process-tracker/src/feature/components/index.ts
@@ -2,4 +2,3 @@ export * from './common';
 export * from './JobExecutionList';
 export * from './JobExecutionDetail';
 export * from './LastExecutionTable';
-export * from './ProgressCell';
